refactor: Store vector summaries in `VectorData.metadata`
