+++ uvdat/core/rest/layer.py
@@ -1,6 +1,5 @@
 from django.db import transaction
 import jsonschema
-from rest_framework.decorators import action
 from rest_framework.response import Response
 from rest_framework.viewsets import ModelViewSet, ReadOnlyModelViewSet
 
@@ -15,11 +14,6 @@ class LayerViewSet(ReadOnlyModelViewSet):
     permission_classes = [GuardianPermission]
     filter_backends = [GuardianFilter]
     lookup_field = 'id'
-
-    @action(detail=True, methods=['get'])
-    def summary(self, request, **kwargs):
-        instance = self.get_object()
-        return Response(instance.get_summary(), status=200)
 
 
 class LayerFrameViewSet(ReadOnlyModelViewSet):
