+++ uvdat/core/tasks/dataset.py
@@ -123,6 +123,7 @@ def convert_dataset(
         # Create vector features after geojson_data may have
         # been altered by create_network or create_source_regions
         create_vector_features(vector_data)
+        vector_data.get_summary()
 
     create_layers_and_frames(dataset, layer_options)
